sval: create cast_rl() which casts a range list to a different type
If the first range list falls outside the list of possible values then it
just returns the whole range for that type.  That's not totally the right
thing to do, but it works for now.

I use this function in smatch_extra.c match_comparison() because I want to
get rid of a place which uses the pre sval version of range lists.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>