sval: handle assignments in smatch extra.
Basically it's what you would expect.  Take the implied min-max of the
right side, cast that to the type of the left side and safe.

The one complication is that we handled equivalent variables as a special
case.  x = y.  If x and y are different types, then should we still count
them as equivalent?  I don't know the answer to that.  For now, I have
said no.

I introduced the types_equiv() function to for test this.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>