sval: extra: absolute: change how assignments are handled
Imagine that we have:
	char x;
	int y;
	...
	y = x;

We know that y is between (-128)-127, but this shouldn't be stored in
smatch_extra.c it should be in smatch_absolute.c.  The difference is that
smatch_extra.c is affects get_implied_value() which is used to detect
array overflows.  Smatch_absolute.c affects get_absolute_min/max().

Also I changed how smatch absolute calculates the min and max.  Using
cast_rl() is more accurate.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>