sval: extra: more fixes to match comparison
This is a fix for the situation where we don't know the value of one side
of a compare.  Before we took the whole range of the promoted type, but
the right thing to do is to take the range of one side of the comparison
and then cast it to the promoted type.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>