sval: hack is so get_absolute_min/max_sval() never fails
Back in the day we could just return whole_range.min/max so these never
failed.  These days we care about the type and there are some times when
we don't know the type so they can fail.  Instead of fixing the callers
I have implemented a temporary hack for now.  Sorry.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>