sval: use typed version of parse_value_ranges() throughout
I made parse_value_ranges_type() fall back to &llong_ctype if we don't know
the type.  Which is possible, and I don't know a better way to handle that.

I removed the old unused parse_value_range().

In smatch_function hooks we sometimes use parse_value_ranges_type() to get
the ranges on the right hand side of the assignment but we should as well
cast the values before setting the left hand state.  I've done that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>