extra: clear hard max if we are falling back to the type max on loops
The problem here is like this:
	if (i = 0; i < some_var; i++) {
		...
	}

If we don't know the value of some_var then the inside the loop the value
is type_max - 1.  The false path is type_max.

The problem is that type_max is just a single value, and we single values
are automatically a hard max.  Which normally works, but in this case we
need to clear the hard max setting.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>