info: dereferences_param: print the dereferences even if it's non-NULL
There was code here to ignore dereferences if we knew already that we are
always passed a valid pointer.

I doesn't hurt anything to print more information and I was got annoyed
that this info was not printed yesterday.  The thing is that the data in
the database is always changing.  I think if you just print as much
information as you know then the results will be more reproducible.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>