db: introduce add_returned_state_callback()
These callbacks will be used to fill the return_states table in the
database.  Everything that wants a return_id will have to be written
like that.

It passes an slist as a parameter.  Use that instead of the cur_slist.
Right now they are equivalent, but later it will change.

I have updated the "return_marker" function to the new callback system.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>