db: call return_states at the end of the function if needed
If we reach the end of a void function that's equivalent to returning.  We
should treat it like a return and call all the callbacks instead of making
it a special case.

Really it should be handled this way in Smatch flow.  But right now I just
want to get this working so I'll look at that later.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>