Revert "constraints: fix bug in remove_from_equiv()"
This reverts commit 7fed101e3bd7acee4aed04c12fc0deb13a71a048.

This causes a seg fault if get_sm_state() returns NULL.  The thing is
that we want to inherit the implications from equivalent states and
so it means that sometimes we will have ->left and ->right where the
sm->left->name is not the same as sm->name.  The way I see it, this
doesn't cause a problem so long as we don't forget this obscure
weirdness.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>