function_hooks: add some calls to __unnullify_path()
The problem is that we call a function:
	some_func(x, y, z);

We want to create an slist for each return path in some_func().  If we were
storing the return value then it wouldn't be a problem.  But since we're
not there could be a return path which doesn't set any states so it gets
ignored.  This is why return_markers exist so that we know when to call
__unnullify_path().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>