implied: introduce overwrite_states_using_pool()
The situation here is that I have a "return foo;".  I know that "foo" can
be 0, -EINVAL, or -ENOMEM.  All the possible values of foo are stored in
the ->possible slist.  I don't care so much about the merged states, only
the original places where the error codes were set.  So I pass the each
of the possible states here to find out what implications are.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>