db: fixup printing return info for difficult to split things
I'm not sure about that this is a complete fix.

When I wrote this I was expecting that we would always some states in
the ->possible list that weren't ->merged.  The reason this happens is
because implied states are merged but we don't store the ->left and ->right
because only part of them are saved...

I'm not sure that smatch_implied.c is doing the right thing here.  Perhaps
the right fix is changing that...  Have to look into that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>