param_limit: don't print empty states
"empty" states represent an impossible condition.  You can have impossible
conditions, but what's happening here is that smatch_implied.c is messing
things up.

It thinks the implied condition is &merged.  Which is true.  But the
->possible list is not set up correctly.

The correct way to fix this is to rewrite this and also change how unmatched
states are handled so cur_slist is correct.  That's a more involved fix so
I'm implementing this temporary hack.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>