type: get_member_type_from_key() should take an expression not a symbol
When I first wrote get member type from key it was only getting the types
for the arguments when a function is declared.  We only have the symbol
there.

Later I used to get the type for arguments passed to functions.  In that
case we need to use the expression otherwise we it passes the same symbol
for both "foo" and "&foo".

This fixes the bug getting the type for "&foo".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>