*new* check_missing_break.c: check for fall through case statements
This was my attempt to find missing break statements.  Falling through is,
of course, allowed by C but so much of the time when people do it, it is
unintentional.

This check tests if a value from the previous statement gets over written
in the next case statement without being used.  I perhaps over thought
things and should just print an error for every time it falls through.

Normally static checker tools let you silence the warning by putting a
/* fall through */ comment on the line before.  Smatch works on the
pre-processed code so it can't see the comments.

It is what it is.

This found 7 kernel bugs and I fixed them all so now it is 100% false
positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>