smatch: introduce expr_to_str_complex()
This version doesn't need a NULL parameter.

As a side note, I don't really care for the _complex() on the end.  It sort
of made sense to talk about complex variables but it doesn't make sense to
talk about complex expressions.

The reason for the split is that if you have a pointer "p" and an offset
"i" and you do "p[i] = 1;" then if either "p" or "i" changes you don't
know what p[i] is.

I originally had the _complex() as default and a _simple() version.  But
then I made _simple() the default and created the _complex() version.

Naming is hard.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>