smatch_comparison: store how two variables compare with each other
The goal here is to keep a record of which variables are larger and which
are smaller, even if we don't know what the values are.  That way we can
say "a - b" is greater than zero if "a" is larger than "b".

This patch just stores the information, it doesn't use it.

Also it only gets the information if there is a compare.  It could use
math operations like divide and subtract as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>