extra, modification_hooks: set "*p" to unknown if we call "function(p)"
This only affects people who are not using the smatch_db.  Before if you
called "foo(&bar)", then it would set "bar" to unknown.  Now if you call
"foo(bar)" it sets "*bar" to unknown.  It does not set "bar->xxx" to
unknown.

It's awkward that smatch_extra.c and smatch_modification_hooks.c both
needed to be modified.  There is some cleanup work here that could be done.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>