math: use smatch_comparison.c for handling subtraction
Basically the idea here is that if we know "a" is larger than "b" then
"a - b" is a number between "1 - a".

The implementation isn't totally right.  There are some issues with
signedness.  But I've been using it for a while and it doesn't seem to
cause many false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>