*new* check_string_len.c: look for sprintf() overflows
This looks for places where sprintf() could lead to an overflow.

This generates about 200 warnings.

There are quite a few false positives.  There were a few serious bugs which
needed to be fixed.  And then there are a bunch of cases where root could
cause an overflow or which I consider sloppy but not worth fixing.

I considered disabling this unless --spammy was used but then I decided
that long term I want to encourage people to use snprintf() anyway.

This check often assumes that INT_MIN is possible, because Smatch is a bit
limited.  It might be a good idea to turn that off.  We'll see.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>