absolute: delete smatch_absolute.c
The hard max/min stuff replaced smatch_absolute.c.

Say we have the following code:

	unsigned char x = z;

We know that since x is a char it can't be more than 255.  That's useful
information, but we shouldn't complain if x is used as an offset into
a 4 element array.

The idea was to store the 255 upper limit in smatch_absolute.c and any
lower limits we could find in smatch_extra.c.

But now we store everything in smatch_extra.c and we set the ->hard_max
flag if we think we're sure about it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>