math: get_implied_min() should be able to return s32min
Before if we saw a s32min, we said that wasn't really an implied min at all
that probably meant the value was unknown.  But actually we want to return
s32min.  It makes get_implied_min() and get_implied_max() match.  It makes
a lot of math work better.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>