sval: only print "s32max" if the value is actually signed
Say if we cast a signed value to unsigned, we still have the value s32max
as an important value in the range list.  But it's confusing to see
s32max in an unsigned variable.  It made me confused.  It's better to just
print the raw value in that case.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>