math: get_implied_max() should be able to return s32max
When I changed get_implied_min() I said that get_implied_max() could return
s32max but it turns out that it couldn't.  It should do.

The reason we want this to work is because there are times when we know
either the upper bound or the lower bound.  With the old code we ended up
discarding both sides when we tried to use the information for math.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>