db: print less --info for a few problematic functions
There are a couple functions where we are tracking a ton of values.  Mostly
these are functions which have some big structs and maybe those structs
have pointers to other structs.  It ends up taking a long time to print
the information on the --info run.

What this does is that when we hit the:

	return ret;

Just print the combined information instead of trying to break it down
and printing a copy of for every value of ret.

It means that instead of 2000 values out six times we just print 2000
values one time.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>