ranges: don't allow inverted ranges
When the min is more than max that's bogus information, don't save it.

Part of the problem why we even have to deal with this is that
smatch_math.c is buggy and incomplete.

I changed alloc_whole_range() as well to handle void types.  Those have
a size of -2 and cause the min to be greater than the max.  It causes an
endless loop.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>