db: take static vs global into consideration wit function pointers
If we're going to save a function pointer to a static function it has to
be done in the same file where the function is defined.

Before it would say the function was called as a pointer when it wasn't
and we'd think all kinds of bogus values passed were passed to it.  This
silences some false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>