param_limit: handle functions where the parameter is not modified
This goes to a lot of effort to handle modified parameters in the right way
but it was ignoring the easier case where the parameter is not modified.

So for a function that does:

int some_func(int x)
{
	return x > 1 && x < 6;
}

It wasn't printing any information for that.  But now it does.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>