db: tools for printing raw SQL instead of human readable info for the db
The plan is that, instead of printing human readable output that no one
actually reads, we can just print SQL.  This will mean we don't need the
smatch_scripts/db/fill_*.pl scripts so it adding new stuff to the database
easier.  It eliminates a layer where bugs can be introduced.

But also, more importantly, once everything is moved to the new format, I
want to change how inline functions are handled.  I'll create a in memory
database and load the call information there and delete it after it has
been used.

I have created the fn_static() function which returns 1 if the current
function being parsed is static or 0 if not.  It's like global_static()
but without the human readable intermediate step.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>