db: move caller_info as close as possible to raw SQL
I can't move caller_info to use raw SQL because I want each call to have
a unique call id.  That lets us group the records for one call together.
The other thing which is a bit tricky is that I want to ignore the
functions which are called over 200 times.  Both of those are easier to
do after we have processed the whole kernel.

But I still want to move it to something closer to SQL because it will
make future work a bit easier.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>