db: make sql_insert_caller_info() take the call expression
Before this took the name of the function and whether it was static or
not.  Now it takes the call expression and figures it out on its own.

The main reason to do this is because I want to use the pointer as a
unique identifier in the In-Memory database.  The other reason to do this
is it makes the code a little cleaner (at the cost of a small slow down).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>