db, flow: parse inline statements, load stuff into the in-memory database
I have introduced two new global variables "__inline_call" and
"__inline_fn".

__inline_call is when you are parsing an inline call.  This is a very
small thing, it's only turned on while we call one one hook.

__inline_fn is turned on when we are processing an inline function.  Some
checks need this like check_locking, which gets confused if we return from
inside an inline statement.

The rule on inline statements is that we inline everything which 10 lines
or less.  But if it's called from inside a function we are already inlining
the I don't inline it.

This patch loads the data into the in-memory database but it doesn't use
it yet.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>