db, flow: make __inline_fn a pointer to the function call instead of an int
Before it was just an on off switch but now when we're processing the
function we can tell where we were called from.  This is also stored in
the in-memory database caller_info table as the function_id already.
(Although really it should probably be renamed to the caller_id).

I will use this later.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>