db: use in-memory database
We've got the in-memory database initialized and loaded with data so now
we can finally use it.  :)

Basically if the function is inlinable() the we prefer use the in-memory
database and return.  Otherwise use the on-disk database.

The one which is a bit different is for caller info where we check that
we are inside an inlined function and we use the call expression saved in
__inline_fn to get the caller_info.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>