db: remove duplicate struct pointers to functions
Gar.  I accidentally committed some temporary code a few minutes ago.  I
actually want to delete this line not comment it out.

The global_variable() function hooks into the same place and prints the
same data.  The global_variable() function is slightly better because it
handles arrays as well.  So I'll keep that version and delete the
GLOBAL_ASSIGNMENT_HOOK.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>