db: use "distinct" to avoid duplicate function pointers
Right now we just bail if there is a duplicate function pointer.  That's
probably not a great idea.  Also we sometimes record a single function
pointer assignment twice so it shows up as a duplicate even though it
isn't one in the source.

Use "select distinct" to make the situation less bad.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>