local_values: store possible values of static data
Static data is normally not passed in as a parameter so it's hard to track
the values of it.

This tries to put it into the database to say that static value "x" is
a number between 1-100.  It looks at how the number was initialized.  Then
if the number gets set inside a function it records that at the end.  At
the end of processing the file it combines all the possible values that
a variable can be and records that in the database.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>