buf_size: handle empty arrays on the end of structs
It's a common idiom to do:

struct my_struct {
	int x, y, z;
	char buf[0];
};

struct my_struct *p = malloc(sizeof(*my_struct) + 100);

That gives you a struct with a dynamically sized array on the end.  Up
to now Smatch hasn't tried to track the size of those arrays.  This patch
adds support for doing that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>