Revert "kernel: remove some unneeded code"
This reverts commit f88989c3a1bff37c787ec02540facce4df0c9651.

These functions *would* be handled by the inline code, but
unfortunately it conflicts with check_err_ptr_deref.c and also
with check_returns.c.

The problem is that when we manually start setting states with
a function call hook or with a return_implies hook then it
conflicts with the information from the database.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>