kernel: handle __ftrace_bad_type()
There isn't actually an __ftrace_bad_type().  GCC can tell at compile
time that it's not used so it doesn't try to link it.  If it were to try
to link then the build would fail.

Treat it like a no return function to discard the states on that path.

This fixes a bunch of false positives.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>