conditions: fix small signedness bug in compares
If I recall correctly, we always get signed comparisons here and manually
change them to SPECIAL_UNSIGNED_ as needed.  In comparisons the type is
promoted to whichever side has more positive bits.  So a comparison between
and int and a unsigned char is signed.  In the original code it would have
been treated as unsigned.

I've updated check_signed.c to match.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>