flow: save and restore the big_expression_stack when processing inlines
This was a simple over sight.  It causes a bug when we look at the big
expression stack to see if a function call is assigned to something.  It
means we query the database twice.

I don't know if this affects any of the shipping Smatch checks but querying
the database twice affects one of my checks that I haven't pushed yet.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>