flow: handle inlines better
Smatch handles inlines in a funny way, because Sparse gives them to us in
a funny way.

Sparse gives us a symbol_list with the functions and global variables in
it.  It doesn't have inline functions on the list.  Before Smatch would
record if it any of the normal functions called an inline function and
process those at the end of the function where they are called.

The problem was that system calls are inline but they are not called by
anything so they weren't processed.

So what I do is I find the last variable in the file.  I take the scope of
it.  I can find the inline functions in that scope list.

It's sort of a hack.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>