locking: take returns values into consideration
There are a lot of places which, for example, take the lock on success
and return zero but don't take the lock on the failure paths.  In the old
code this would generate a warning but now it shouldn't.

I use the broken-apart return value so if we get "return err;" it looks
at the individual paths which merge into "err" instead of taking them all
together.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>