db: always split the return values if we're not using --info
This code is to handle the case where we do:

		err = -ENOMEM;
		goto err_out;
	...
		err = -EINVAL;
		goto err_out;
err_out:
	unlock();
	return err;

It's often better to try separate each return path and figure out the
implications of each.  The problem is that if it's a big function with a
lot of states and there are a bunch of return paths then sometimes it's
thousands of states and it's just too much information to print.  In those
cases we just bail out.

But if we're not printing information then it doesn't take too much time to
separate the states.

The only thing this affects currently is check_locking.c.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>