rosenberg: remove pahole dependency
The main problem is that pahole was not finding some structs which had
holes.

With this change, I've made it so it prints where the hole is as well.  If
there is more than one hole then it only prints the first one.

There are still some false positives.  For example, if there is a nested
struct, it only checks that the outside struct has been initialized.  So
if the inside struct has holes and it has been initialized but the outside
one has not then it prints a false positive warning.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>