ranges: an empty range is not a whole range
In the original code the reason I treated an empty range like a whole range
was because I thought it was an impossible condition so it meant the value
could be anything.

But it turns out there are valid reasons to have an empty range.  For
example, in param_filter an empty range means none of the original values
are preserved.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>