flow: look at static and not-static local inline functions
The way that Smatch processes inline functions is a hack.  We take the last
symbol in the file and look at the functions with the same scope.  But
sometimes the last symbol is static and sometimes it's not.  We actually
want to look at both static and non-static inline functions.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>