db: introduce call_results_to_rl()
I would like to store relationships in the database instead of just
numbers.  For example the return value from copy_from_user() is between
zero and the size parameter.  I want to store this in the database:

OLD:  0-u64max
NEW:  0-<2>

The "<2>" here means parameter 2.

This patch doesn't change what is loaded in the database, so it doesn't
change how Smatch works.  It just changes how the database info would be
parsed if "0-<2>" were stored there.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>