sval: fix sval_too_high()
The problem is that if there was a number that too high so it got cast to
negative then it was less than the max.  I've changed it to just compare
the .uvalue values which is simpler.

I've changed sval_is_negative() in the same way because I think it makes
it easier to read.  That part of the change is just a cleanup.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>